Manager registers ALL controllers for restic too
